bookmanager 수정
1. Book 클래스 따로 뺌
2. comparator랑 binary search 수정
3. 주석 추가
